+++ src/Presentation/Nop.Web/Administration/Controllers/NewsController.cs
@@ -257,6 +257,8 @@ namespace Nop.Admin.Controllers
                 return AccessDeniedView();
 
             var comment = _customerContentService.GetCustomerContentById(id);
+            if (comment == null)
+                throw new ArgumentException("No comment found with the specified id");
             _customerContentService.DeleteCustomerContent(comment);
 
             return Comments(filterByNewsItemId, command);
